requirements.txt
flask
pymysql
python-dotenv
requests
psutil
